Extract repeater directory imports into web/js/ui/repeater-query.js
The przemienniki.net / RepeaterBook filter modal, its dictionary fetch, the
query and the result insertion now live in one module owning the endpoint
config, the per-source dictionary promises and the active source. The
API-base meta tag resolution moves with it.

Also prunes the imports and element refs ui.js no longer uses.
